requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
blis==0.7.9
brotlipy==0.7.0
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
comtypes==1.1.10
confection==0.0.3
constantly==15.1.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.7
Cython
cytoolz==0.11.0
dnspython==2.3.0
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1-py3-none-any.whl
essential-generators==1.0
et-xmlfile==1.1.0
fastapi==0.88.0
ffmpy==0.3.0
fire==0.5.0
flake8
Flask
Flask-Cors==3.0.10
flatbuffers==22.11.23
fonttools==4.25.0
frozendict==2.3.4
frozenlist==1.3.3
gast==0.4.0
gensim
gitdb==4.0.10
GitPython==3.1.31
glob2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==3.12.0
greenlet
grpcio==1.50.0
h11==0.12.0
h5py
HeapDict
holoviews
html5lib==1.1
httpcore==0.15.0
httpx==0.23.1
huggingface-hub==0.14.1
idna==2.10
imgaug==0.4.0
imutils==0.5.4
inflection==0.5.1
keras==2.11.0
Keras-Applications==1.0.8
langcodes==3.3.0
libclang==14.0.6
linkify-it-py==1.0.3
llvmlite==0.38.0
markdown-it-py==2.1.0
mccabe==0.6.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mock
mpmath==1.2.1
msgpack
multidict==6.0.3
multipledispatch
multitasking==0.0.11
munkres==1.1.4
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
ninja==1.11.1
nltk==3.8.1
numpy==1.21.5
numpydoc==1.4.0
oauthlib==3.2.2
olefile
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.8.3
packaging==21.3
pandas==1.4.4
pathy==0.10.1
patsy==0.5.2
pep8==1.7.1
Pillow==9.2.0
preshed==3.0.8
protobuf==3.19.6
psutil==5.9.0
py==1.11.0
pyasn1-modules==0.2.8
pybboxes==0.1.6
pyclipper==1.3.0.post4
pycosat==0.6.3
pycryptodome==3.16.0
pydantic==1.10.2
PyDispatcher==2.0.5
pydub==0.25.1
pyls-spyder==0.4.0
pymongo==4.3.3
pyparsing==2.4.7
pytest==7.1.2
python-bidi==0.4.2
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
pywin32==302
PyYAML==6.0
queuelib==1.5.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
roboflow==1.0.9
rsa==4.9
sahi==0.11.14
scikit-image
scikit-learn
scipy
Scrapy==2.6.2
seaborn==0.11.2
sentry-sdk==1.23.1
shapely==2.0.1
smmap==5.0.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
starlette==0.22.0
statsmodels==0.13.2
tables==3.6.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminaltables==3.1.10
thinc==8.1.6
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
typer==0.7.0
uc-micro-py==1.0.1
ultralytics==8.0.105
uvicorn==0.20.0
validators==0.20.0
wasabi==0.10.1
webencodings==0.5.1
websockets==10.4
wget==3.2
win-unicode-console==0.5
wincertstore==0.2
yarl==1.8.2
zict==2.1.0